精彩视频加上淡出

+++ sh-engine/src/main/java/com/sh/engine/plugin/BatchSegVideoMergePlugin.java
@@ -1,15 +1,13 @@
 package com.sh.engine.plugin;
 
 import com.google.common.collect.Lists;
-import com.sh.config.manager.ConfigFetcher;
-import com.sh.config.model.config.StreamerConfig;
-import com.sh.engine.base.StreamerInfoHolder;
+import com.sh.config.utils.EnvUtil;
+import com.sh.config.utils.VideoFileUtils;
 import com.sh.engine.service.VideoMergeService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.FileFilterUtils;
-import org.apache.commons.lang3.BooleanUtils;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Resource;
@@ -34,12 +32,6 @@ public class BatchSegVideoMergePlugin implements VideoProcessPlugin {
     @Override
     public boolean process(String recordPath) {
         // 只有录像才能进行合并
-        String name = StreamerInfoHolder.getCurStreamerName();
-        StreamerConfig streamerConfig = ConfigFetcher.getStreamerInfoByName(name);
-        if (BooleanUtils.isNotFalse(streamerConfig.isRecordWhenOnline())) {
-            return false;
-        }
-
         Collection<File> tsFiles = FileUtils.listFiles(new File(recordPath), FileFilterUtils.suffixFileFilter("ts"), null);
         if (CollectionUtils.isEmpty(tsFiles)) {
             return true;
@@ -54,7 +46,7 @@ public class BatchSegVideoMergePlugin implements VideoProcessPlugin {
         for (List<Integer> batchIndexes : Lists.partition(segIndexes, BATCH_RECORD_TS_COUNT)) {
             File targetMergedVideo = new File(recordPath, "P" + videoIndex + ".mp4");
             List<String> segNames = batchIndexes.stream()
-                    .map(i -> new File(recordPath, "seg-" + i + ".ts"))
+                    .map(i -> new File(recordPath, VideoFileUtils.genSegName(i)))
                     .map(File::getAbsolutePath)
                     .collect(Collectors.toList());
             boolean success;
@@ -63,10 +55,10 @@ public class BatchSegVideoMergePlugin implements VideoProcessPlugin {
                 success = true;
             } else {
                 // 合并视频
-                success = videoMergeService.merge(segNames, targetMergedVideo);
+                success = videoMergeService.concat(segNames, targetMergedVideo);
             }
 
-            if (success) {
+            if (success && EnvUtil.isProd()) {
                 deleteSegs(segNames);
             }
             videoIndex++;
